[WIP] Make Nix work again (#29)
* Make Nix work again

* Remove Pandas as a dependency from Nix build

* Fix issue on Travis CI with boto config

Try and fix an issue on Travis CI with `ModuleNotFoundError: No module
named 'google_compute_engine'` error. This link seems to have an idea
about what to do,

https://github.com/travis-ci/travis-ci/issues/7940

Adding Pandas back to the buildInputs as it wasn't anything to do with
the error.

* Try to stop timeout on Travis CI

Travis CI may be failing because of the length of the log. Unsure
about this, but redirecting stderr to /dev/null since most of the Nix
blurb goes to stderr.

* Adding travis_wait to avoid timeout on Travis CI

After adding "2> /dev/null" to the first nix-shell means that no
output occurs for at least 10 minutes, which make Travis fail. Adding
travis_wait stops that fail.

* Remove Pandas again as it's taking ages to build

* Direct all output to /dev/null

* Use --no-build-output in place of > /dev/null

* Trying to make Travis work

* Splitting out numpy, scipy and pandas in Travis CI

* Add travis_wait to failing scipy build

* Use travis_wait with Pandas build

* Remove pandas from build

* Remove pandas as a requirement in setup.py

* Allow build output from main nix-shell command

* Try again

* Trying stuff

* Build stuff separately

* Fix travis CI

* Fix Nix issues with nixpkgs version

* Run scipy build for longer

* Fix broken command for Travis CI

* more changes

* Add travis_wait to pymatgen build

* Split out pymatgen packages

* Split out matplotlib

* Make Pandas not print everything and use travis_wait

* Try with older version of nixpkgs

* See Pandas build output

* Just build pandas using unstable Nixpkgs

* Try again with version 18.09 of Nix

* Install everything and see if Travis CI works

* Add test back in

* Check if nix-shell can run solo

See if nix-shell can all be done in one command now.